app/requirements.txt
streamlit
crewai
python-dotenv
langchain_groq